requirements.txt
asgiref==3.6.0
boto3==1.26.106
botocore==1.29.106
Django==4.1.7
django-storages==1.13.2
djangorestframework==3.14.0
jmespath==1.0.1
Pillow==9.5.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
